refactor(PaymentType): 클래스의 책임 분산
