+++ src/main/java/domain/Table.java
@@ -1,5 +1,6 @@
 package domain;
 
+import java.util.Collections;
 import java.util.HashMap;
 import java.util.Map;
 
@@ -48,7 +49,7 @@ public class Table {
     }
 
     public Map<Menu, Integer> getOrderHistory() {
-        return orderHistory;
+        return Collections.unmodifiableMap(orderHistory);
     }
 
     public int getNumber() {
